Add files via upload
a C program that reads letters from a text file named words.txt into a two-dimensional array. In each column of the array, there is an English word. For example, the word “house” is written in the first column and “choir” is written in the last column. The program will take a number input and copy the corresponding word from the two-dim array into a one-dim array. Then the program will display the contents of the one-dim array.